Python translation of a part of RGFA
This is a squash of the commits in the Python branch.

Author of a large part of the code is Tim Weber,
written under my supervision (Giorgio Gonnella).

As he forgot to add his E-mail address,
I had to squash all commits, otherwise all rebases
had to be doen manually each time.